// client/src/api/models/accounts/account-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * 家庭记账系统的 RESTful API，支持用户认证、人员管理、账户管理、交易记录和统计。 支持银行、微信、支付宝等多种账号，每种账号下可有多张卡或子账户。 角色权限： - 管理员：访问所有数据，管理用户、账号、子账户。 - 普通用户：仅访问通过 user_sub_accounts 关联的子账户和自己的交易/统计。 
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Account } from './account';

/**
 * 
 * @export
 * @interface AccountType
 */
export interface AccountType {
    /**
     * 
     * @type {number}
     * @memberof AccountType
     */
    'typeId'?: number;
    /**
     * 
     * @type {string}
     * @memberof AccountType
     */
    'typeName'?: string;
    /**
     * 
     * @type {Array<Account>}
     * @memberof AccountType
     */
    'accounts'?: Array<Account>;
}